requirements.txt
streamlit==1.30.0
pandas==2.1.4
openpyxl==3.1.2
bcrypt==4.1.2
langgraph>=1.0.0
python-dateutil==2.8.2
